feat: added css in app and condicional class

+++ app/front-end/src/pages/approvRecipe.tsx
@@ -5,6 +5,7 @@ import Header from "../components/header";
 import pastaSoTastyContext from '../context/context';
 import IRecipe from '../interface/IRecipe';
 import ImageLoader from '../helpers/imageLoader';
+import '../styles/pages/approvRecipe.css';
 
 const AprovarReceitas: React.FC = () => {
   const [recipes, setRecipes] = useState<IRecipe[]>([]);
@@ -31,45 +32,64 @@ const AprovarReceitas: React.FC = () => {
   }, [id, fullName]);
 
   const handleAccept = async (id: number) => {
-    await approveRecipe(id);
-    const response = await fetchAllPedingRecipes();
-    setRecipes(prev => prev = response);
-    navigate('/aprovar-receitas');
+    const confirmMessage = "Você realmente deseja aceitar esta receita?";
+    const userConfirmed = window.confirm(confirmMessage);
+    if (userConfirmed) {
+      await approveRecipe(id);
+      const response = await fetchAllPedingRecipes();
+      setRecipes(prev => prev = response);
+      navigate('/aprovar-receitas');
+    }
   };
 
   const handleDelete = async (id: number) => {
-    await deleteRecipe(id);
-    const response = await fetchAllPedingRecipes();
-    setRecipes(prev => prev = response);
-    navigate('/aprovar-receitas');
+    const confirmMessage = "Você realmente deseja excluir esta receita?";
+    const userConfirmed = window.confirm(confirmMessage);
+    if (userConfirmed) {
+      await deleteRecipe(id);
+      const response = await fetchAllPedingRecipes();
+      setRecipes(prev => prev = response);
+      navigate('/aprovar-receitas');
+    }
   };
 
   return (
     <div>
       <Header isUserLoggedIn={logged} fullName={fullName}>
       </Header >
-      <h1>Receitas aguardando aprovação</h1>
-      {recipes.length > 0 ? (
-        recipes.map((recipe: IRecipe) => (
-          <div key={recipe.recipe_name}>
-            <Link to={`/receita/${recipe.id}-${recipe.recipe_name.split(' ').join('-')}`}
-            >
-              <h3>{recipe.recipe_name}</h3>
-            </Link>
-            {
-              recipe.recipe_photo.toLowerCase().startsWith('http') ? (
-                <img src={recipe.recipe_photo} alt={recipe.recipe_name} width="200" height="150" />
-              ) : (
-                <ImageLoader photo={recipe.recipe_photo} alt={recipe.recipe_name} />
-              )
-            }
-            <button onClick={() => handleAccept(recipe.id)}>Aceitar</button>
-            <button onClick={() => handleDelete(recipe.id)}>Negar</button>
+      <div className="home-recipes">
+        <h1>Receitas aguardando aprovação</h1>
+        {recipes.length > 0 ? (
+          <div className="recipe-list">
+            {recipes.map((recipe: IRecipe) => (
+              <div key={recipe.recipe_name} className={`recipe-card${recipes.length === 1 ? ' single-card' : ''}`}>
+                {
+                  recipe.recipe_photo.toLowerCase().startsWith('http') ? (
+                    <img src={recipe.recipe_photo} alt={recipe.recipe_name} width="200" height="150" />
+                  ) : (
+                    <ImageLoader photo={recipe.recipe_photo} alt={recipe.recipe_name} width={"200"} height={"150"} />
+                  )
+                }
+                <div className='approv-card-name'>
+                  <Link to={`/receita/${recipe.id}-${recipe.recipe_name.split(' ').join('-')}`}
+                    className='link-name'
+                  >
+                    <div className='approv-card-h3'>
+                      <h3>{recipe.recipe_name}</h3>
+                    </div>
+                  </Link>
+                </div>
+                <div>
+                  <button onClick={() => handleAccept(recipe.id)} className='btn-visu'>Aceitar</button>
+                  <button onClick={() => handleDelete(recipe.id)} className='btn-visu'>Negar</button>
+                </div>
+              </div>
+            ))}
           </div>
-        ))
-      ) : (
-        <p>Nenhuma receita encontrada</p>
-      )}
+        ) : (
+          <p>Nenhuma receita encontrada</p>
+        )}
+      </div>
     </div >
   );
 };
